Rename test files to match source file names and fix test method names
- Renamed test files to follow [SourceFileName]Tests.cs pattern:
  - SemanticKernelApiTests.cs → SemanticKernelAPITests.cs (match API capitalization)
  - ShellTests.cs → ShellViewModelTests.cs
  - BackendTests.cs → BackendServiceTests.cs
  - ProblemVMTests.cs → ProblemViewModelTests.cs
  - LockTest.cs → LockTests.cs
  - Merged ShellUITests.cs into StoryModelTests.cs (they test StoryModel)

- Fixed test method names to follow MethodName_Scenario_ExpectedResult pattern
- Added test naming conventions to CLAUDE.md
- Added branch management guidelines to CLAUDE.md

Part of issue #1069: Achieve 100% API and Outline Service Test Coverage
