--- StoryCADTests/ShellUITests.cs
@@ -1,70 +1,0 @@
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using StoryCAD.Models;
-using StoryCAD.ViewModels;
-
-namespace StoryCADTests
-{
-    [TestClass]
-    public class ShellUITests
-    {
-        [TestMethod]
-        public void TrashView_SeparateFromMainViews_IsImplemented()
-        {
-            // This test verifies that TrashView is a separate collection
-            // from ExplorerView and NarratorView
-            
-            // Arrange
-            var storyModel = new StoryModel();
-            
-            // Act & Assert
-            Assert.IsNotNull(storyModel.TrashView, "TrashView collection should be initialized");
-            Assert.IsNotNull(storyModel.ExplorerView, "ExplorerView collection should be initialized");
-            Assert.IsNotNull(storyModel.NarratorView, "NarratorView collection should be initialized");
-            
-            // Verify they are separate collections
-            Assert.AreNotSame(storyModel.TrashView, storyModel.ExplorerView, "TrashView should be separate from ExplorerView");
-            Assert.AreNotSame(storyModel.TrashView, storyModel.NarratorView, "TrashView should be separate from NarratorView");
-        }
-
-        [TestMethod]
-        public void TrashView_CollectionChanges_SetChangedFlag()
-        {
-            // Arrange
-            var storyModel = new StoryModel();
-            var deletedScene = new SceneModel("Deleted Scene", storyModel, null);
-            var deletedItem = new StoryNodeItem(deletedScene, null);
-            
-            // Reset changed flag
-            storyModel.Changed = false;
-            
-            // Act
-            storyModel.TrashView.Add(deletedItem);
-            
-            // Assert
-            Assert.IsTrue(storyModel.Changed, "Adding to TrashView should set Changed flag");
-        }
-
-        [TestMethod]
-        public void DragAndDrop_Constraints_Documentation()
-        {
-            // This test documents the expected drag-and-drop behavior:
-            // 
-            // 1. Main NavigationTree (Explorer/Narrator views):
-            //    - CanDragItems="True"
-            //    - AllowDrop="True"
-            //    - CanReorderItems="True"
-            //
-            // 2. TrashView TreeView:
-            //    - CanDragItems="False" (cannot drag items out of trash)
-            //    - AllowDrop="False" (cannot drop items into trash via drag)
-            //    - CanReorderItems="False" (cannot reorder within trash)
-            //
-            // 3. Cross-TreeView Operations:
-            //    - Not possible because TreeViews are separate controls
-            //    - Move to trash: Use context menu "Delete" command
-            //    - Restore from trash: Use context menu "Restore" command
-            
-            Assert.IsTrue(true, "Drag-and-drop constraints are enforced through XAML properties");
-        }
-    }
-}
